+++ api/src/application/film_record/command/interface/film_record_command_application.py
@@ -1,5 +1,7 @@
+# 標準ライブラリ
 from abc import ABC, abstractmethod
 
+# 独自ライブラリ
 from base_model.film_record.create.film_record_create_model import FilmRecordCreateModel
 from src.domain.film_record import FilmRecordEntity, FilmRecordIdObject, IFilmRecordRepository
