+++ api/src/application/film_record/query/film_record_query_application.py
@@ -2,16 +2,14 @@
 from logging import getLogger
 
 # 独自ライブラリ
-from src.application.film_record.query.film_record_query_domain_service import ImplFilmRecordQueryDomainService
-from src.application.film_record.query.interface.film_record_query_application_interface import (
-    IFilmRecordQueryApplication,
-)
-from src.application.film_record.query.interface.film_record_query_domain_service import IFilmRecordQueryDomainService
+from src.application.film_record.query.interface import IFilmRecordQueryApplication, IFilmRecordQueryDomainService
 from src.domain.film_record import FilmRecordEntity, FilmRecordIdObject, IFilmRecordRepository
 from src.domain.film_record.film.genre import IFilmGenreRepository
-from src.domain.film_record.film.series.film_series_repository import IFilmSeriesRepository
+from src.domain.film_record.film.series import IFilmSeriesRepository
 from src.domain.film_record.watch_history import IFilmWatchHistoryRepository
-from src.domain.film_record.watch_medium.watch_medium_repository import IWatchMediumRepository
+from src.domain.film_record.watch_medium import IWatchMediumRepository
+
+from .film_record_query_domain_service import ImplFilmRecordQueryDomainService
 
 
 class ImplFilmRecordQueryApplication(IFilmRecordQueryApplication):
