+++ api/src/presentation/containers.py
@@ -1,9 +1,10 @@
+# 外部ライブラリ
 from dependency_injector import containers
 from dependency_injector.providers import Factory
-from src.application.film_record.command.interface.film_record_command_application import IFilmRecordCommandApplication
-from src.application.film_record.query.interface.film_record_query_application_interface import (
-    IFilmRecordQueryApplication,
-)
+
+# 独自ライブラリ
+from src.application.film_record.command.interface import IFilmRecordCommandApplication
+from src.application.film_record.query.interface import IFilmRecordQueryApplication
 from src.application.jumanpp.jumanpp_application import JumanppApplication
 from src.presentation.film_record.command.film_record_command_container import FilmRecordCommandContainer
 from src.presentation.film_record.query.film_record_query_container import FilmRecordQueryContainer
